Move/refactor PT helpers to hybrid.reference.pt submodule
