+++ examples/parallel-tempering-qpu.py
@@ -28,6 +28,9 @@ import neal
 import dimod
 import hybrid
 
+from hybrid.reference.pt import FixedConstantTemperatureSampler
+from hybrid.reference.pt import SweepDownReplicasSwap
+
 
 # load a problem
 problem = sys.argv[1]
@@ -41,62 +44,6 @@ print("BQM: {} nodes, {} edges, {:.2f}% density".format(n, m, d))
 
 
 # PT workflow: temperature/beta is a property of a branch
-
-class FixedTemperatureSampler(hybrid.Runnable, hybrid.traits.SISO):
-    """PT propagate/update step.
-
-    On each call, run fixed temperature (~`1/beta.state`) simulated annealing
-    for `num_sweeps` (seeded by input sample(s)), effectively producing a new
-    state by sampling from a Boltzmann distribution at the given temperature.
-    """
-
-    def __init__(self, beta, num_sweeps=10000, **runopts):
-        super(FixedTemperatureSampler, self).__init__(**runopts)
-        self.beta = beta
-        self.num_sweeps = num_sweeps
-
-    def next(self, state, **runopts):
-        new_samples = neal.SimulatedAnnealingSampler().sample(
-            state.problem, initial_states=state.samples,
-            beta_range=(self.beta, self.beta), beta_schedule_type='linear',
-            num_sweeps=self.num_sweeps).aggregate()
-
-        return state.updated(samples=new_samples)
-
-
-class SwapReplicasSweepDown(hybrid.Runnable, hybrid.traits.MIMO):
-    """PT swap replicas step.
-
-    On each call, sweep through and probabilistically swap all adjacent pairs
-    of replicas (input states).
-    """
-
-    def __init__(self, betas, **runopts):
-        super(SwapReplicasSweepDown, self).__init__(**runopts)
-        self.betas = betas
-
-    def swap_pair(self, states, i, j):
-        """One pair of states (i, j) probabilistic swap."""
-
-        beta_diff = self.betas[i] - self.betas[j]
-        energy_diff = states[i].samples.first.energy - states[j].samples.first.energy
-
-        # since `min(1, math.exp(beta_diff * energy_diff))` can overflow,
-        # we need to move `min` under `exp`
-        w = math.exp(min(0, beta_diff * energy_diff))
-        p = random.uniform(0, 1)
-        if w > p:
-            # swap samples for replicas i and j
-            states[i], states[j] = states[j], states[i]
-
-        return states
-
-    def next(self, states, **runopts):
-        for i in range(len(states) - 1):
-            states = self.swap_pair(states, i, i + 1)
-
-        return states
-
 
 n_sweeps = 10000
 n_replicas = 10
@@ -118,10 +65,10 @@ qpu = hybrid.IdentityDecomposer() | hybrid.QPUSubproblemAutoEmbeddingSampler() |
 # use QPU as the hottest temperature sampler and `n_replicas-1` fixed-temperature-samplers
 update = hybrid.Branches(
     qpu,
-    *[FixedTemperatureSampler(beta, num_sweeps=n_sweeps) for beta in betas[1:]])
+    *[FixedConstantTemperatureSampler(beta, num_sweeps=n_sweeps) for beta in betas[1:]])
 
 # swap step is `n_replicas-1` pairwise potential swaps
-swap = SwapReplicasSweepDown(betas)
+swap = SweepDownReplicasSwap(betas)
 
 # we'll run update/swap sequence for `n_iterations`
 workflow = hybrid.Loop(update | swap, max_iter=n_iterations) \
