+++ examples/parallel-tempering.py
@@ -26,6 +26,8 @@ import neal
 import dimod
 import hybrid
 
+from hybrid.reference.pt import FixedStateTemperatureSampler, RandomPairSampleSwap
+
 
 # load a problem
 problem = sys.argv[1]
@@ -39,58 +41,6 @@ print("BQM: {} nodes, {} edges, {:.2f}% density".format(n, m, d))
 
 
 # PT workflow, one approach: store temperatures/betas in states
-
-class FixedTemperatureSampler(hybrid.Runnable, hybrid.traits.SISO):
-    """PT propagate/update step.
-
-    On each call, run fixed temperature (~`1/beta.state`) simulated annealing
-    for `num_sweeps` (seeded by input sample(s)), effectively producing a new
-    state by sampling from a Boltzmann distribution at the given temperature.
-    """
-
-    def __init__(self, num_sweeps=10000, **runopts):
-        super(FixedTemperatureSampler, self).__init__(**runopts)
-        self.num_sweeps = num_sweeps
-
-    def next(self, state, **runopts):
-        new_samples = neal.SimulatedAnnealingSampler().sample(
-            state.problem, initial_states=state.samples,
-            beta_range=(state.beta, state.beta), beta_schedule_type='linear',
-            num_sweeps=self.num_sweeps).aggregate()
-
-        return state.updated(samples=new_samples)
-
-
-class SwapReplicas(hybrid.Runnable, hybrid.traits.MIMO):
-    """PT swap replicas step.
-
-    On each call, choose a random input state (replica), and probabilistically
-    accept a swap with the adjacent state (replica).
-    """
-
-    def swap_pair(self, states, i, j):
-        """One pair of states' (i, j) samples probabilistic swap."""
-
-        s_i, s_j = states[i], states[j]
-        beta_diff = s_i.beta - s_j.beta
-        energy_diff = s_i.samples.first.energy - s_j.samples.first.energy
-
-        # since `min(1, math.exp(beta_diff * energy_diff))` can overflow,
-        # we need to move `min` under `exp`
-        w = math.exp(min(0, beta_diff * energy_diff))
-        p = random.uniform(0, 1)
-        if w > p:
-            # swap samples for replicas i and j
-            states[i].samples, states[j].samples = s_j.samples, s_i.samples
-
-        return states
-
-    def next(self, states, **runopts):
-        i = random.choice(range(len(states) - 1))
-        j = i + 1
-
-        return self.swap_pair(states, i, j)
-
 
 n_sweeps = 10000
 n_replicas = 10
@@ -111,8 +61,8 @@ replicas = hybrid.States(*[state.updated(beta=b) for b in betas])
 
 # run replicas update/swap for n_iterations
 # (after each update/sampling step, do n_replicas-1 swap operations)
-update = hybrid.Map(FixedTemperatureSampler(num_sweeps=n_sweeps))
-swap = hybrid.Loop(SwapReplicas(), max_iter=n_random_swaps)
+update = hybrid.Map(FixedStateTemperatureSampler(num_sweeps=n_sweeps))
+swap = hybrid.Loop(RandomPairSampleSwap(), max_iter=n_random_swaps)
 workflow = hybrid.Loop(update | swap, max_iter=n_iterations) \
          | hybrid.MergeSamples(aggregate=True)
