+++ hybrid/reference/__init__.py
@@ -12,4 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from hybrid.reference.kerberos import KerberosSampler
+from hybrid.reference.kerberos import *
+from hybrid.reference.pt import *
